pyproject.toml
[tool.poetry]
name = "GeVhoo_diff_calc"
version = "0.1.5"
description = "Difference calculator"
authors = ["GeVhoo <gvahrushev@mail.ru>"]
repository = "https://github.com/GeVhoo/python-project-lvl2"
packages = [
	{ include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
pytest = "^5.3"
pytest-cov = "^2.8"
coverage = "^5.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
